// work/AzureDevOps/angular/app/src/app/modules/produtos/sub-modules/e-contrato/core/models/contratos/credor.model.ts
import { Contato } from "../common/contato.model";
import { Documento } from "../common/documento.model";
import { Endereco } from "../common/endereco.model";

export class Credor {
    empresaId?: number;
    agenteFinanceiro: number;
    codigoAgenteFinanceiro: string;
    nomeAgenteFinanceiro: string;
    documento: Documento;
    endereco: Endereco;
    contato: Contato;

    constructor() {
        this.documento = new Documento();
        this.endereco = new Endereco();
        this.contato = new Contato();
    }
}